Add album-download command
Generalize the existing user-download command and add a new one.
