+++ .phpcs.xml
@@ -19,4 +19,10 @@
         <exclude name="SlevomatCodingStandard.Functions.UnusedParameter.UnusedParameter" />
         <exclude name="SlevomatCodingStandard.TypeHints.TypeHintDeclaration.UselessDocComment" />
     </rule>
+    <arg name="tab-width" value="4" />
+    <rule ref="Generic.WhiteSpace.ScopeIndent">
+        <properties>
+            <property name="tabIndent" value="true" />
+        </properties>
+    </rule>
 </ruleset>
