+++ i18n/en.json
@@ -32,7 +32,7 @@
   "unable-to-download": "Unable to download: $1 $2",
   "added-checksum": "Added checksum machine tag: $1 $2",
 
-  "command-download-desc": "Download photos and compile metadata files according to a template.",
+  "command-download-user-desc": "Download a user's photos and compile metadata files according to a template.",
   "option-dest-desc": "The directory to download files to. Will be created if it doesn't exist.",
   "option-template-desc": "The template to use for the downloaded files. One of: $1",
   "option-userid-desc": "The ID of the user who's photos to return. The default value of 'me' will return the authorized user's photos.",
@@ -41,6 +41,10 @@
   "option-min-date-taken-desc": "Only download photos taken on or after this date-time. Use ISO8601 date format.",
   "option-max-date-taken-desc": "Only download photos taken on or before this date-time. Use ISO8601 date format.",
   "option-privacy-desc": "When downloading your own photos, restrict to only: 1 = public; 2 = friends; 3 = family; 4 = friends & family; 5 = private.",
+  "retrieving-photo-metadata": "Retrieving photo metadata",
   "compiling-output-files": "Compiling output files.",
-  "downloaded-saved-to": "Photos downloaded to: $1"
+  "downloaded-saved-to": "Photos downloaded to: $1",
+
+  "command-download-album-desc": "Download an album's photos and compile metadata files according to a template.",
+  "option-album-desc": "The ID of the album to download. Get it from the URL of the album page on Flickr."
 }
